mew room static response
refs #2

this adds in a potential chance of a room being the "mew" room for
the wifi plaza. this is the authoratative source since the lobby
CHANKEY is always taken over the web one. i have not yet figured
out the correct "empty" response to send without the game either
crashing, or asking for it infinitely. that is tbd.

NOTE: i do plan on opening up a follow up pr once i have all the fields
figured out, as well as the correct empty response for peerchat.
if you'd like another room in the meantime, @InternalLoss you know
how to reach me :)