rewrite to python3, fully support mew&commands
at this point miniircd-peerchat has been completely rewritten to
being a more fully fleshed python3 app that is hopefully easier
to extend in the future, easier to understand how it works for folks,
and can actually handle dependencies we want to take on without a
huge pain in ensuring the correct things are installed. since we've
got poetry now.

the miniircd-peerchat server now also fully supports the mew room,
as well as letting the web server (in this case pkmn-classic's C#
code) determine what action we should take.

we also now support UTM commands, and have added in some protections
for clients/server alike to not accept obviously bogus values. right
now the 'deep' parsing (not just the length checks) are non-blocking.
we just log errors if we can't parse them. i've tested every message
in terms of pokemon platinum/heartgold/soulsilver for the plaza rooms
that i can find. so hopefully we can enforce this sometime soon.